+++ app/components/flat_pack/sidebar/group/component.rb
@@ -170,6 +170,7 @@ module FlatPack
           classes(
             "flex-1",
             "text-left",
+            "fp-sidebar-label",
             ("sr-only" if @collapsed)
           )
         end